fix(notebook): align namespace browser contracts
