+++ src/main/notebook/runtime-service.architecture.test.ts
@@ -240,6 +240,7 @@ describe('Notebook runtime facade architecture', () => {
         'finishCodeCell',
         'getActiveNotebookSessions',
         'getSessionReference',
+        'inspectNamespace',
         'inspectPackages',
         'isDefaultEnvRecoveryBlocked',
         'isPrefixLiveUnconfirmed',
